% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_config.R
\name{add_run_configuration}
\alias{add_run_configuration}
\title{add_run_configuration}
\usage{
add_run_configuration(config, filename, refname, start, end)
}
\arguments{
\item{config}{}

\item{filename}{}

\item{refname}{}

\item{start}{}

\item{end}{}
}
\description{
add_run_configuration
}
\author{
Martin Vincent
}

