% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{chain.info.epiG}
\alias{chain.info.epiG}
\title{chain.info}
\usage{
\method{chain.info}{epiG}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
??
}
\description{
chain.info
}
\author{
Martin Vincent
}

