% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{chain_info.epiG}
\alias{chain_info.epiG}
\title{Information about chains}
\usage{
\method{chain_info}{epiG}(object, ...)
}
\arguments{
\item{object}{epiG model}

\item{...}{ignored}
}
\value{
A data.frame with \code{nchain(object)} rows.
}
\description{
Rerive information about chains
}
\examples{
#TODO examples chain_info
}
\author{
Martin Vincent
}

