\name{compute_chunk_positions}
\alias{compute_chunk_positions}
\title{Compute chunk positions}
\usage{
  compute_chunk_positions(filename, refname, start, end,
    chunk_size)
}
\arguments{
  \item{filename}{}

  \item{refname}{}

  \item{start}{}

  \item{end}{}

  \item{chunk_size}{}
}
\value{
  chunk positions
}
\description{
  Compute chunk positions
}
\author{
  martin
}

