% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{coverage}
\alias{coverage}
\title{coverage}
\usage{
coverage(object, pos = NULL, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}
}
\value{
??
}
\description{
coverage
}
\author{
Martin Vincent
}

