\name{create_error_distributions}
\alias{create_error_distributions}
\title{create_error_distributions}
\usage{
  create_error_distributions(bisulfite_rate = 0.94,
    bisulfite_inap_rate = 0.06)
}
\arguments{
  \item{bisulfite_rate}{}

  \item{bisulfite_inap_rate}{}
}
\value{
  bisulfite model
}
\description{
  create_error_distributions
}
\author{
  Martin Vincent
}

