\name{create_genotype_prior}
\alias{create_genotype_prior}
\title{create_genotype_prior}
\usage{
  create_genotype_prior(scale = 0.5, R = exp(6 * scale),
    RA = exp(5 * scale), A = exp(4 * scale),
    AA = exp(3 * scale), RAA = exp(2 * scale),
    AAA = exp(scale), RAAA = 1)
}
\arguments{
  \item{scale}{}

  \item{R}{}

  \item{RA}{}

  \item{A}{}

  \item{AA}{}

  \item{RAA}{}

  \item{AAA}{}

  \item{RAAA}{}
}
\value{
  prior
}
\description{
  create_genotype_prior
}
\author{
  martin
}

