\name{create_genotype_prior_alt}
\alias{create_genotype_prior_alt}
\title{create_genotype_prior_alt}
\usage{
  create_genotype_prior_alt(scale = 0.5,
    R = exp(11 * scale), RA = exp(10 * scale),
    A = exp(9 * scale), RB = exp(8 * scale),
    AB = exp(7 * scale), B = exp(6 * scale),
    RAB = exp(5 * scale), RBB = exp(4 * scale),
    ABB = exp(3 * scale), RABB = exp(2 * scale), BB = 1)
}
\arguments{
  \item{scale}{}

  \item{R}{}

  \item{RA}{}

  \item{A}{}

  \item{RB}{}

  \item{AB}{}

  \item{B}{}

  \item{RAB}{}

  \item{RBB}{}

  \item{ABB}{}

  \item{RABB}{}

  \item{BB}{}
}
\value{
  prior
}
\description{
  create_genotype_prior_alt
}
\author{
  martin
}

