% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{end.epiG}
\alias{end.epiG}
\title{end position}
\usage{
\method{end}{epiG}(object, ...)
}
\arguments{
\item{object}{fitted model (epiG object)}

\item{...}{}
}
\value{
end position (bp) of region contined in model
}
\description{
end
}
\examples{
#TODO
}
\author{
Martin Vincent
}

