% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{end}
\alias{end}
\title{End position}
\usage{
end(object, ...)
}
\arguments{
\item{object}{an epiG model}

\item{...}{additonal arguments}
}
\value{
the last position in model
}
\description{
Return the last position in the model
}
\author{
Martin Vincent
}

