\name{epiG.chunks}
\alias{epiG.chunks}
\title{epiG.chunks}
\usage{
  epiG.chunks(filename, refGenom_filename,
    altGenom_filename, refname, chunks_start, chunks_end,
    max_threads = 8L, config = msbs.standard.config)
}
\arguments{
  \item{filename}{}

  \item{refGenom_filename}{}

  \item{altGenom_filename}{}

  \item{refname}{}

  \item{chunks_start}{}

  \item{chunks_end}{}

  \item{max_threads}{}

  \item{config}{}
}
\value{
  fitted models
}
\description{
  epiG.chunks
}
\author{
  martin
}

