% Generated by roxygen2 (4.0.1): do not edit by hand
\name{epiG}
\alias{epiG}
\title{epiG}
\usage{
epiG(filename, refname, start, end, max_threads = 8L, config,
  refGenom_filename = config$ref.filename,
  altGenom_filename = config$alt.filename)
}
\arguments{
\item{filename}{}

\item{refGenom_filename}{}

\item{altGenom_filename}{}

\item{refname}{}

\item{start}{}

\item{end}{}

\item{max_threads}{}

\item{config}{}
}
\value{
fitted model
}
\description{
epiG
}
\author{
martin
}

