% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_haplo_fit.R
\name{epiG_chunks}
\alias{epiG_chunks}
\title{epiG_chunks}
\usage{
epiG_chunks(configs, max_threads = 8L)
}
\arguments{
\item{max_threads}{}
}
\value{
fitted models
}
\description{
epiG_chunks
}
\author{
Martin Vincent
}

