% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_config.R
\name{epiG_config}
\alias{epiG_config}
\title{epiG_config}
\usage{
epiG_config(model, ref_file, alt_file, max_iterations = 1e+05, ref_prior = 1
  - 1e-04, min_overlap_length = 50, min_CG_count = 1, min_HCGD_count = 0,
  min_DGCH_count = 0, margin = 5, max_stages = 1,
  structual_prior_scale = 1, chunk_size = 5000, chunk_method = "reads",
  reads_hard_limit = 7500, quality_threshold = 0.02,
  use_paired_reads = FALSE, split_mode = FALSE, verbose = TRUE)
}
\arguments{
\item{model}{}

\item{ref_file}{}

\item{alt_file}{}

\item{max_iterations}{}

\item{ref_prior}{}

\item{min_overlap_length}{}

\item{min_CG_count}{}

\item{min_HCGD_count}{}

\item{min_DGCH_count}{}

\item{margin}{}

\item{max_stages}{}

\item{structual_prior_scale}{}

\item{chunk_size}{}

\item{chunk_method}{}

\item{reads_hard_limit}{}

\item{quality_threshold}{}

\item{use_paired_reads}{}

\item{split_mode}{}

\item{verbose}{}
}
\value{
...
}
\description{
epiG_config
}
\author{
Martin Vincent
}

