% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_fetch.R
\name{fetch.reads.raw}
\alias{fetch.reads.raw}
\title{fetch.reads.raw}
\usage{
fetch.reads.raw(filename, refname, start, end)
}
\arguments{
\item{filename}{}

\item{refname}{}

\item{start}{}

\item{end}{}
}
\value{
reads
}
\description{
fetch.reads.raw
}
\author{
martin
}

