% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_fetch.R
\name{fetch_read_count}
\alias{fetch_read_count}
\title{fetch_reads_info}
\usage{
fetch_read_count(filename, refname, start, end)
}
\arguments{
\item{filename}{}

\item{refname}{}

\item{start}{}

\item{end}{}
}
\value{
info
}
\description{
fetch_reads_info
}
\author{
martin
}

