% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{fetch_read_info}
\alias{fetch_read_info}
\title{Information About Reads}
\usage{
fetch_read_info(file, refname, start, end)
}
\arguments{
\item{file}{path to bam file}

\item{refname}{reference name}

\item{start}{start of region}

\item{end}{end of region}
}
\value{
data.frame with information abut reads. Columns:
\itemize{
 \item[\code{name}]{name of read}
 \item[\code{start}]{start position of read}
 \item[\code{end}]{end position of read}
 \item[\code{length}]{length of read}
}
}
\description{
Fetch information about reads overlapping the specified region
}
\examples{
# Retrieve paths to raw data files
bam_file <- system.file("extdata", "GNAS_small.bam", package="epiG")

fetch_read_info(bam_file, "chr20", 57400000, 57400000 + 100)

}
\author{
Martin Vincent
}

