% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_fetch.R
\name{fetch_reads_raw}
\alias{fetch_reads_raw}
\title{fetch_reads_raw}
\usage{
fetch_reads_raw(filename, refname, start, end)
}
\arguments{
\item{filename}{}

\item{refname}{}

\item{start}{}

\item{end}{}
}
\value{
reads
}
\description{
fetch_reads_raw
}
\author{
Martin Vincent
}

