% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{fetch_ref}
\alias{fetch_ref}
\title{Fetch reference genom}
\usage{
fetch_ref(object)
}
\arguments{
\item{object}{epiG epigenotype model}
}
\value{
epiG epigenotype model
}
\description{
Load reference genom and include it in epiG object.
}
\author{
Martin Vincent
}

