% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{file_info}
\alias{file_info}
\title{Fetch information about bam file}
\usage{
file_info(file)
}
\arguments{
\item{file}{path to bam file}
}
\description{
Fetch information about bam file
}
\author{
Martin Vincent
}

