% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{genotype}
\alias{genotype}
\title{genotype}
\usage{
genotype(object, pos, remove.meth, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{remove.meth}{}

\item{...}{}
}
\value{
??
}
\description{
genotype
}
\author{
Martin Vincent
}

