\name{haplo.epiG}
\alias{haplo.epiG}
\title{Haplo information}
\usage{
  haplo.epiG(model, threshold = 5)
}
\arguments{
  \item{model}{epiG model}

  \item{threshold}{minimum number of overlaping bases}
}
\value{
  data.frame with haplotype information
}
\description{
  Haplo information
}
\author{
  Martin Vincent
}

