% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{header_info}
\alias{header_info}
\title{Fetch Bam Header}
\usage{
header_info(file)
}
\arguments{
\item{file}{path to bam file}
}
\value{
a list of refnames and lengths associated with the file
}
\description{
Load bam file header
}
\examples{
# Retrieve paths to raw data files
bam_file <- system.file("extdata", "GNAS_small.bam", package="epiG")

header_info(bam_file)
}
\author{
Martin Vincent
}

