% Generated by roxygen2 (4.0.1): do not edit by hand
\name{length.epiG}
\alias{length.epiG}
\title{Length of model in base pairs}
\usage{
\method{length}{epiG}(x)
}
\arguments{
\item{object}{}
}
\value{
Length of model in base pairs
}
\description{
Length of model in base pairs
}
\author{
Martin Vincent
}

