% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_locate.R
\name{locate.mutations}
\alias{locate.mutations}
\title{find positions of mutations}
\usage{
locate.mutations(object)
}
\arguments{
\item{object}{}
}
\value{
??
}
\description{
find positions of mutations
}
\author{
Martin Vincent
}

