% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_C}
\alias{locate_C}
\title{locate C positions}
\usage{
locate_C(object)
}
\arguments{
\item{object}{epiG object}
}
\value{
positions of C in object
}
\description{
locate C positions
}
\examples{
todo

}
\author{
Martin Vincent
}

