% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_CG}
\alias{locate_CG}
\title{Locate CpG}
\usage{
locate_CG(object)
}
\arguments{
\item{object}{}
}
\value{
??
}
\description{
Locate CpG positions
}
\examples{
#TODO examle locate_CG
}
\author{
Martin Vincent
}

