% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_locate.R
\name{locate_DGCH}
\alias{locate_DGCH}
\title{locate_DGCH}
\usage{
locate_DGCH(object)
}
\arguments{
\item{object}{}
}
\value{
??
}
\description{
locate DGCH (isolated GpC) positions
}
\author{
Martin Vincent
}

