% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_GC}
\alias{locate_GC}
\title{Locate GpC}
\usage{
locate_GC(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
a vector of GpC positions
}
\description{
Locate GpC positions in the refrence genom
}
\examples{
#TODO example
}
\author{
Martin Vincent
}

