% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_HCGD}
\alias{locate_HCGD}
\title{Locate HCGD}
\usage{
locate_HCGD(object)
}
\arguments{
\item{object}{}
}
\value{
??
}
\description{
locate HCGD (isolated CpG) positions
}
\examples{
#TODO
}
\author{
Martin Vincent
}

