% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_mismatch}
\alias{locate_mismatch}
\title{Locate Mismatches}
\usage{
locate_mismatch(object)
}
\arguments{
\item{object}{}
}
\value{
postions of mismatches
}
\description{
Locate positions where at least one chain has a genotype not matching with the reference.
}
\examples{
#TODO
}
\author{
Martin Vincent
}

