% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_locate.R
\name{locate_nonref}
\alias{locate_nonref}
\title{locate_nonref}
\usage{
locate_nonref(object)
}
\arguments{
\item{object}{}
}
\value{
??
}
\description{
find positions of mutations
}
\author{
Martin Vincent
}

