% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{loglike.epiG}
\alias{loglike.epiG}
\title{genotype
codeing C = 1, G = 2, A = 3, T = 4}
\usage{
\method{genotype}{epiG}(object, pos, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}

\item{remove.meth}{}
}
\value{
??
}
\description{
genotype
codeing C = 1, G = 2, A = 3, T = 4
}
\author{
Martin Vincent
}

