% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{loglike}
\alias{loglike}
\title{genotype}
\usage{
loglike(object, pos, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}
}
\value{
??
}
\description{
genotype
}
\author{
Martin Vincent
}

