% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchain.epiG}
\alias{nchain.epiG}
\title{Number of chains}
\usage{
\method{nchain}{epiG}(object, ...)
}
\arguments{
\item{object}{epiG model}

\item{...}{ignored}
}
\value{
number of chains in the model
}
\description{
Number of chains in the model
}
\examples{
#TODO examples nchain
}
\author{
Martin Vincent
}

