% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchain}
\alias{nchain}
\title{Number of chains}
\usage{
nchain(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
number of chains in the model
}
\description{
Number of chains in the model
}
\examples{
#TODO examples nchain
}
\author{
Martin Vincent
}

