% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchunks.epiG}
\alias{nchunks.epiG}
\title{Number of chunks}
\usage{
\method{nchunks}{epiG}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
??
}
\description{
Number of chunks
}
\author{
Martin Vincent
}

