% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchunks}
\alias{nchunks}
\title{Number of chunks}
\usage{
nchunks(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
the number of chunks in the epiG object
}
\description{
Number of chunks in the epiG object
}
\examples{
data(example)

nchunks(fit)
}
\author{
Martin Vincent
}

