% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-config.R
\name{NOMeSeq}
\alias{NOMeSeq}
\title{Create a NOMe-sequencing conversion model}
\usage{
NOMeSeq(bisulphite_rate = 0.95, bisulphite_inap_rate = 0.05, Lmax = 110,
  ...)
}
\arguments{
\item{bisulphite_rate}{bisulphite conversion rate (numeric in the range (0, 1])}

\item{bisulphite_inap_rate}{bisulphite inappropriate conversion rate (numeric in the range (0, 1])}

\item{Lmax}{maximal read length (integer)}
}
\value{
an epiG conversion model
}
\description{
Create a NOMe-sequencing conversion model
}
\author{
Martin Vincent
}

