% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{nread.epiG}
\alias{nread.epiG}
\title{Number of reads in model}
\usage{
\method{nread}{epiG}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
??
}
\description{
Number of reads in model
}
\author{
Martin Vincent
}

