% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nread}
\alias{nread}
\title{Number of reads in model}
\usage{
nread(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
number of reads contined in model
}
\description{
Number of reads in model
}
\examples{
data(example)

nread(fit)
}
\author{
Martin Vincent
}

