% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{position.info}
\alias{position.info}
\title{position.info}
\usage{
position.info(object, pos, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}
}
\value{
??
}
\description{
position.info
}
\author{
Martin Vincent
}

