% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{position_info.epiG}
\alias{position_info.epiG}
\title{Position Info}
\usage{
\method{position_info}{epiG}(object, pos, ...)
}
\arguments{
\item{object}{epiG model}

\item{pos}{position}

\item{...}{ignored}
}
\value{
??
}
\description{
Retriev information about the estimated epi-genotype at a given position in the model
}
\examples{
#TODO examples position_info
}
\author{
Martin Vincent
}

