% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{position_info}
\alias{position_info}
\title{Position Info}
\usage{
position_info(object, pos)
}
\arguments{
\item{object}{epiG model}

\item{pos}{position}
}
\value{
??
}
\description{
Retriev information about the estimated epi-genotype at a given position in the model
}
\author{
Martin Vincent
}

