% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG.config}
\alias{print.epiG.config}
\title{Print Information About an epiG Configuration}
\usage{
\method{print}{epiG.config}(x, ...)
}
\arguments{
\item{x}{epiG configuration}

\item{...}{ignored}
}
\description{
Print Information About an epiG Configuration
}
\examples{
#TODO examples print.epiG.config
}
\author{
Martin Vincent
}

