% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG.model}
\alias{print.epiG.model}
\title{print model}
\usage{
\method{print}{epiG.model}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
print model
}
\examples{
#TODO examples print.epiG.model 
}
\author{
Martin Vincent
}

