% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG}
\alias{print.epiG}
\title{print}
\usage{
\method{print}{epiG}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
print
}
\examples{
#TODO examples print
}
\author{
Martin Vincent
}

