% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG.reads}
\alias{print.epiG.reads}
\title{Print Information About an epiG Reads Object}
\usage{
\method{print}{epiG.reads}(x, ...)
}
\arguments{
\item{x}{epiG reads object}

\item{...}{ignored}
}
\description{
Print information about an epiG reads object
}
\examples{
#TODO examples print.epiG.reads 
}
\author{
Martin Vincent
}

