% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{read.info.epiG_reads}
\alias{read.info.epiG_reads}
\title{read.info.epiG_reads}
\usage{
\method{read.info}{epiG_reads}(object, inc.symbols = FALSE, ...)
}
\arguments{
\item{object}{}

\item{inc.symbols}{}

\item{...}{}
}
\description{
read.info.epiG_reads
}
\author{
martin
}

