% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_depth.epiG}
\alias{read_depth.epiG}
\title{Read depth}
\usage{
\method{read_depth}{epiG}(object, pos = NULL, ...)
}
\arguments{
\item{object}{epiG model}

\item{pos}{position}

\item{...}{ignored}
}
\value{
??
}
\description{
Retrive the read depth at a given position in the model
}
\examples{
#TODO examples read_depth
}
\author{
Martin Vincent
}

