% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_depth}
\alias{read_depth}
\title{Read depth}
\usage{
read_depth(object, pos = NULL, ...)
}
\arguments{
\item{object}{an epiG model}

\item{pos}{position}

\item{...}{additonal arguments}
}
\value{
??
#TODO return read_depth
}
\description{
Exatract the read depth
}
\author{
Martin Vincent
}

