% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA File}
\usage{
read_fasta(file, refname, start, len, offset = 0)
}
\arguments{
\item{file}{path to fasta file}

\item{refname}{reference name}

\item{start}{start of region}

\item{len}{length of region}

\item{offset}{file offset (position of first base in file)}
}
\value{
TODO
}
\description{
Load a raw FASTA file
}
\examples{
#TODO
}
\author{
Martin Vincent
}

