% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_info.epiG}
\alias{read_info.epiG}
\title{Information about reads}
\usage{
\method{read_info}{epiG}(object, inc.symbols = FALSE, ...)
}
\arguments{
\item{object}{epiG model}

\item{inc.symbols}{if \code{TRUE} each line in the returned data.frame will correspond to one nuleobase, if \code{FALSE} one read.}

\item{...}{ignored}
}
\value{
??
}
\description{
Retrive information about reads in the model
}
\author{
Martin Vincent
}

