% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_info.epiG_reads}
\alias{read_info.epiG_reads}
\title{read_info.epiG_reads}
\usage{
\method{read_info}{epiG_reads}(object, inc.symbols = FALSE, ...)
}
\arguments{
\item{object}{}

\item{inc.symbols}{}

\item{...}{}
}
\description{
read_info.epiG_reads
}
\author{
Martin Vincent
}

