% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_info}
\alias{read_info}
\title{Information about reads}
\usage{
read_info(object, ...)
}
\arguments{
\item{object}{epiG model}

\item{...}{additonal arguments}
}
\value{
??
}
\description{
Retrive information about reads in the model
}
\author{
Martin Vincent
}

