% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-config.R
\name{set_run_configuration}
\alias{set_run_configuration}
\title{Run Configuration}
\usage{
set_run_configuration(config, filename, refname, start, end)
}
\arguments{
\item{config}{an epiG configuration}

\item{filename}{bam file (path to .bam file)}

\item{refname}{reference name}

\item{start}{start position of region to processes}

\item{end}{end position of region to processes}
}
\value{
an epiG configuration
}
\description{
set run configuration
}
\author{
Martin Vincent
}

