% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{start.epiG}
\alias{start.epiG}
\title{Start position}
\usage{
\method{start}{epiG}(x, ...)
}
\arguments{
\item{x}{an epiG model}

\item{...}{ignored}
}
\value{
the first position in the model
}
\description{
Return the first position in the model
}
\examples{
data(example)

start(fit)
}
\author{
Martin Vincent
}

