% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{start}
\alias{start}
\title{start position}
\usage{
start(object, ...)
}
\arguments{
\item{object}{an epiG model}

\item{...}{ignored}
}
\value{
??
}
\description{
start position
}
\author{
Martin Vincent
}

