% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subregion.epiG}
\alias{subregion.epiG}
\title{subregion}
\usage{
\method{subregion}{epiG}(object, start, end, chop.reads = TRUE, ...)
}
\arguments{
\item{object}{}

\item{start}{}

\item{end}{}

\item{chop.reads}{}

\item{...}{}
}
\value{
??
}
\description{
subregion
}
\author{
martin
}

