% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{subregion}
\alias{subregion}
\title{subregion}
\usage{
subregion(object, start, end, chop.reads = FALSE, ...)
}
\arguments{
\item{object}{}

\item{start}{}

\item{end}{}

\item{chop.reads}{}

\item{...}{}
}
\value{
??
}
\description{
subregion
}
\author{
Martin Vincent
}

