% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_locate.R
\name{vector.search}
\alias{vector.search}
\title{vector search}
\usage{
vector.search(pattern, text)
}
\arguments{
\item{pattern}{}

\item{text}{}
}
\value{
??
}
\description{
vector search
}
\author{
Martin Vincent
}

