% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{vector_search}
\alias{vector_search}
\title{Pattern Search}
\usage{
vector_search(pattern, x)
}
\arguments{
\item{pattern}{integer vector}

\item{x}{integer vector to search in}
}
\value{
position of pattern in x
}
\description{
Search for pattern in integer vector
}
\examples{
#TODO
}
\author{
Martin Vincent
}

