% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_config.R
\name{auto_config}
\alias{auto_config}
\title{auto_config}
\usage{
auto_config(ref_file, alt_file, NOMEseq = FALSE, bam_file, chr = NULL,
  start = NULL, end = NULL, use_paired_reads = NULL, min_overlap = NULL,
  ...)
}
\arguments{
\item{ref_file}{}

\item{alt_file}{}

\item{NOMEseq}{}

\item{bam_file}{}

\item{chr}{}

\item{start}{}

\item{end}{}

\item{use_paired_reads}{}

\item{min_overlap}{}

\item{...}{}
}
\description{
auto_config
}
\author{
martin
}

