% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-config.R
\name{auto_config}
\alias{auto_config}
\title{Create Standard Configuration}
\usage{
auto_config(ref_file, alt_file, bam_file, chr, start, end, seq_type = "BSeq",
  use_paired_reads = NULL, ...)
}
\arguments{
\item{ref_file}{genome reference file (path to .fa file)}

\item{alt_file}{alternative nucleotide file (path to .fa file)}

\item{bam_file}{bam file (path to .bam file)}

\item{chr}{reference name}

\item{start}{start position of region to processes}

\item{end}{end position of region to processes}

\item{seq_type}{sequencing type ("BSeq" for bisulphite sequencing, "NOMeSeq" for NOMe sequencing)}

\item{use_paired_reads}{should pair information be used (TRUE/FALSE or NULL, if NULL then paired information is used if pairs are present in bam file)}

\item{...}{additional arguments (overrides default values)}
}
\value{
An epiG configuration
}
\description{
Create a epiG configuration with standard parameters
}
\examples{
#TODO
}
\author{
Martin Vincent
}

