% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{chain_info.epiG}
\alias{chain_info.epiG}
\title{chain_info}
\usage{
\method{chain_info}{epiG}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
??
}
\description{
chain_info
}
\author{
Martin Vincent
}

