% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{chain_info}
\alias{chain_info}
\title{Information about chains}
\usage{
chain_info(object, ...)
}
\arguments{
\item{object}{epiG model}

\item{...}{additonal arguments}
}
\value{
A data.frame with \code{nchain(object)} rows.
}
\description{
Rerive information about chains
}
\author{
Martin Vincent
}

