% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_config.R
\name{create_bisulfite_model}
\alias{create_bisulfite_model}
\title{create_bisulfite_model}
\usage{
create_bisulfite_model(bisulfite_rate, bisulfite_inap_rate, Lmax)
}
\arguments{
\item{bisulfite_inap_rate}{}

\item{Lmax}{}

\item{bisulfite_rates}{}
}
\value{
...
}
\description{
create_bisulfite_model
}
\author{
Martin Vincent
}

