\name{create_bisulfite_model}
\alias{create_bisulfite_model}
\title{create_bisulfite_model}
\usage{
  create_bisulfite_model(bisulfite_rates = 0.94,
    bisulfite_inap_rate = 0.06, lambda = 0.1, Lmax = 100)
}
\arguments{
  \item{bisulfite_rates}{}

  \item{bisulfite_inap_rate}{}

  \item{lambda}{}

  \item{Lmax}{}
}
\value{
  ...
}
\description{
  create_bisulfite_model
}
\author{
  Martin Vincent
}

