% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_config.R
\name{create_error_distributions}
\alias{create_error_distributions}
\title{create_error_distributions}
\usage{
create_error_distributions(bisulfite_rate, bisulfite_inap_rate)
}
\arguments{
\item{bisulfite_rate}{}

\item{bisulfite_inap_rate}{}
}
\value{
bisulfite model
}
\description{
create_error_distributions
}
\author{
Martin Vincent
}

