% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{end.epiG}
\alias{end.epiG}
\title{End position}
\usage{
\method{end}{epiG}(x, ...)
}
\arguments{
\item{x}{an epiG model}

\item{...}{ignored}
}
\value{
the last position in model
}
\description{
Return the last position in the model
}
\examples{
data(example)

end(fit)
}
\author{
Martin Vincent
}

