% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_config.R
\name{epiG.algorithm.config}
\alias{epiG.algorithm.config}
\title{Create a epiG configuration}
\usage{
epiG.algorithm.config(model, ref.file, alt.file, max_iterations = 1e+05,
  ref_prior = 0.99, min_overlap_length = 50, min_CG_count = 1,
  min_HCGD_count = 0, min_DGCH_count = 0, margin = 5, max_stages = 25,
  structual_prior_scale = 1, chunk_size = 5000, chunk_method = "reads",
  reads_hard_limit = 7500, use_paired_reads = FALSE, split_mode = FALSE,
  verbose = TRUE)
}
\arguments{
\item{model}{}

\item{ref.file}{}

\item{alt.file}{}

\item{max_iterations}{}

\item{ref_prior}{}

\item{min_overlap_length}{}

\item{chunk_size}{}

\item{chunk_method}{}

\item{reads_hard_limit}{}

\item{verbose}{}

\item{prior}{}

\item{sequence_quality_adjust}{}

\item{haplo_prior}{modeled using a geometric distribution}
}
\value{
epiG configuration
}
\description{
Create a epiG configuration
}
\author{
martin
}

