\name{epiG.chunks}
\alias{epiG.chunks}
\title{epiG.chunks}
\usage{
  epiG.chunks(filename, refnames, chunks_start, chunks_end,
    max_threads = 8L, config,
    refGenom_filename = config$ref.filename,
    altGenom_filename = config$alt.filename)
}
\arguments{
  \item{filename}{}

  \item{refGenom_filename}{}

  \item{altGenom_filename}{}

  \item{refnames}{}

  \item{chunks_start}{}

  \item{chunks_end}{}

  \item{max_threads}{}

  \item{config}{}
}
\value{
  fitted models
}
\description{
  epiG.chunks
}
\author{
  martin
}

