% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_haplo_fit.R
\name{epiG.chunks}
\alias{epiG.chunks}
\title{epiG.chunks}
\usage{
epiG.chunks(configs, max_threads = 8L)
}
\arguments{
\item{max_threads}{}

\item{filename}{}

\item{refGenom_filename}{}

\item{altGenom_filename}{}

\item{refnames}{}

\item{chunks_start}{}

\item{chunks_end}{}

\item{config}{}
}
\value{
fitted models
}
\description{
epiG.chunks
}
\author{
martin
}

