\name{epiG}
\alias{epiG}
\title{epiG}
\usage{
  epiG(filename, refGenom_filename, altGenom_filename,
    refname, start, end, max_threads = 8L,
    config = msbs.standard.config)
}
\arguments{
  \item{filename}{}

  \item{refGenom_filename}{}

  \item{altGenom_filename}{}

  \item{refname}{}

  \item{start}{}

  \item{end}{}

  \item{max_threads}{}

  \item{config}{}
}
\value{
  fitted model
}
\description{
  epiG
}
\author{
  martin
}

