% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_haplo_fit.R
\name{epiG}
\alias{epiG}
\title{Fit an epiG model}
\usage{
epiG(config, max_threads = 8L)
}
\arguments{
\item{config}{}

\item{max_threads}{}

\item{filename}{BAM file with aligned and ordered reads}

\item{altGenom_filename}{}

\item{refname}{}

\item{start}{}

\item{end}{}
}
\value{
fitted model
}
\description{
Fit an epiG model
}
\author{
Martin Vincent
}

