% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fit.R
\name{epiG}
\alias{epiG}
\title{Fit an epiG epigenotype model}
\usage{
epiG(config, max_threads = 2L)
}
\arguments{
\item{config}{epiG configuration}

\item{max_threads}{maximal number of threads to use}
}
\value{
fitted model
}
\description{
Fit an epiG epigenotype model
}
\examples{
library(epiG)

# Retrieve paths to raw data files
bam_file <- system.file("extdata", "GNAS_small.bam", package="epiG")
ref_file <- system.file("extdata", "hg19_GNAS.fa", package="epiG")
alt_file <- system.file("extdata", "dbsnp_135.hg19_GNAS.fa", package="epiG")

# Specify region
chr <- "chr20"
start <- 57400000 
end <- 57400000 + 1000

# Build epiG configuration
config <- auto_config(
		bam_file = bam_file,
		ref_file = ref_file,
		alt_file = alt_file,
		chr = chr,
		start = start,
		end = end,
# If ref_file and alt_file contains the entire chromosome this is not needed
		ref_offset = 57380000, 
		alt_offset = 57380000)

# Run epiG
fit <- epiG(max_threads = 2, config = config)

# Information about fitted model
fit 

# Information about haplotype chains
chain_info(fit)

}
\author{
Martin Vincent
}

