% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fit.R
\name{epiG_chunks}
\alias{epiG_chunks}
\title{Fit epiG epigenotype models}
\usage{
epiG_chunks(configs, max_threads = 8L)
}
\arguments{
\item{configs}{list of epiG configurations}

\item{max_threads}{maximal number of threads to use}
}
\value{
list of fitted models
}
\description{
Fit an epiG epigenotype model for each config in the list configs.
}
\author{
Martin Vincent
}

