% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-config.R
\name{epiG_config}
\alias{epiG_config}
\title{Create an epiG Configuration}
\usage{
epiG_config(ref_file = NULL, alt_file = NULL, ref_offset = 1,
  alt_offset = 1, model, min_overlap, min_CG, min_HCGD, min_DGCH,
  ref_prior = 1 - 1e-04, structual_prior = 1, quality_threshold = 0.02,
  margin = 5, chunk_method = "reads", chunk_size = 5000,
  hard_limit = 6000, paired_reads, max_iterations = 1e+05, max_stages = 1,
  verbose = TRUE, ...)
}
\arguments{
\item{ref_file}{genome reference file (path to .fa file)}

\item{alt_file}{alternative nucleotide file (path to .fa file)}

\item{ref_offset}{ref file offset (usually ref_offset = 1)}

\item{alt_offset}{alt file offset (usually alt_offset = 1)}

\item{model}{conversion model}

\item{min_overlap}{minimum overlapping length}

\item{min_CG}{minimum overlapping CG positions}

\item{min_HCGD}{minimum overlapping HCGD positions}

\item{min_DGCH}{minimum overlapping DGCH positions}

\item{ref_prior}{genotype prior parameter}

\item{structual_prior}{structural prior scaling}

\item{quality_threshold}{discard reads with mean epsilon quality higher than quality_threshold}

\item{margin}{cut off margin}

\item{chunk_method}{Method used to split region into chunks ('none' only one chunk, 'reads' chunks of approximately chunk_size reads, 'bases' chunks of chunk_size bases)}

\item{chunk_size}{chunk size}

\item{hard_limit}{maximal number of reads loaded per chunk (reads not loaded will be completely ignored)}

\item{paired_reads}{use pair information (reads with the same name in the bam file is paired and will be forced into the same haplotype chain)}

\item{max_iterations}{maximal number of iterations}

\item{max_stages}{experimental stage optimization (if <= 1 then stage optimization is off)}

\item{verbose}{show information while running}

\item{...}{ignored}
}
\value{
an epiG configuration
}
\description{
Create a custom epiG configuration
}
\examples{

# Retrieve paths to raw data files
ref_file <- system.file("extdata", "hg19_GNAS.fa", package="epiG")
alt_file <- system.file("extdata", "dbsnp_135.hg19_GNAS.fa", package="epiG")

config <- epiG_config(

	ref_file = ref_file,
	alt_file = alt_file,
	ref_offset = 57380000,
	alt_offset = 57380000,

	model = BSeq(),

	min_overlap = 80,
	min_CG = 0,
	min_HCGD = 0,
	min_DGCH = 0,

	ref_prior = 0.999,
	structual_prior = 1,
	margin = 5,
	quality_threshold = 0.020,

 chunk_method = "reads",
	chunk_size = 8000,
	hard_limit = 10000,

	paired_reads = TRUE,

	max_iterations = 1e5,

	verbose = TRUE
	)

config # print a summary of the configuration

# Specify region and bam file
bam_file <- system.file("extdata", "GNAS_small.bam", package="epiG")

chr <- "chr20"
start <- 57400000
end <- 57400000 + 1000

config <- set_run_configuration(config, bam_file, chr, start, end)

# Run epiG
fit <- epiG(max_threads = 2, config = config)

fit # print a summary of inferred model
}
\author{
Martin Vincent
}

