% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{fetch_alt}
\alias{fetch_alt}
\title{Fetch Alternative Nucleotides}
\usage{
fetch_alt(object)
}
\arguments{
\item{object}{}
}
\value{
epiG model
}
\description{
Load alternative nucleotides and include it in epiG object.
}
\author{
Martin Vincent
}

