% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{fetch_ref}
\alias{fetch_ref}
\title{Fetch Reference Genom}
\usage{
fetch_ref(object)
}
\arguments{
\item{object}{epiG epigenotype model}
}
\value{
epiG epigenotype model with reference genome included
}
\description{
Load reference genome and include it in epiG object.
}
\examples{
#TODO
}
\author{
Martin Vincent
}

