% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{file_info}
\alias{file_info}
\title{Fetch Bam File Information}
\usage{
file_info(file)
}
\arguments{
\item{file}{path to bam file}
}
\value{
a data.frame with the following columns:
\itemize{
 \item[\code{ref}]{refname}
 \item[\code{length}]{length of ref}
 \item[\code{nreds}]{number of reads assigned to refname}
 \item[\code{mean_read_length}]{the mean read length of reads assigned to refname}
}
}
\description{
Fetch information about bam file
}
\examples{
# Retrieve paths to raw data files
bam_file <- system.file("extdata", "GNAS_small.bam", package="epiG")

file_info(bam_file)
}
\author{
Martin Vincent
}

