\name{genotype.epiG}
\alias{genotype.epiG}
\title{genotype
codeing C = 1, G = 2, A = 3, T = 4}
\usage{
  \method{genotype}{epiG} (object, pos,
    remove.meth = FALSE, ...)
}
\arguments{
  \item{object}{}

  \item{pos}{}

  \item{remove.meth}{}

  \item{...}{}
}
\value{
  ??
}
\description{
  genotype codeing C = 1, G = 2, A = 3, T = 4
}
\author{
  Martin Vincent
}

