% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{header_info}
\alias{header_info}
\title{Fetch Bam Header}
\usage{
header_info(file)
}
\arguments{
\item{file}{path to bam file}
}
\value{
TODO
}
\description{
Load bam file header
}
\examples{
#TODO
}
\author{
Martin Vincent
}

