% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{header_info}
\alias{header_info}
\title{Fetch header of bam file}
\usage{
header_info(file)
}
\arguments{
\item{file}{path to bam file}
}
\description{
Fetch header of bam file
}
\author{
Martin Vincent
}

