% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{length.epiG}
\alias{length.epiG}
\title{Length of model}
\usage{
\method{length}{epiG}(x)
}
\arguments{
\item{x}{}
}
\value{
Length of model in base pairs
}
\description{
Length of model in base pairs
}
\examples{
#TODO
}
\author{
Martin Vincent
}

