% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{load_reads}
\alias{load_reads}
\title{Load Reads}
\usage{
load_reads(file, refname, start, end, quality_threshold = 1,
  raw_quality_scores = FALSE)
}
\arguments{
\item{file}{path to bam file}

\item{refname}{reference name}

\item{start}{start of region}

\item{end}{end of region}

\item{quality_threshold}{quality threshold}

\item{raw_quality_scores}{if TRUE raw quality score will be returned}
}
\value{
a list with the following entries:
\itemize{
 \item[\code{reads}]{a list of reads (each read represented by a vector of bases)}
 \item[\code{quality}]{a list quality scores}
 \item[\code{positions}]{a vector of the start positions of the reads}
 \item[\code{lengths}]{a vector of the lengths of the reads}
 \item[\code{names}]{a vector of the names of the reads}
}
}
\description{
Load the reads overlapping the specified region
}
\examples{
# Retrieve paths to raw data files
bam_file <- system.file("extdata", "GNAS_small.bam", package="epiG")

info <- load_reads(bam_file, "chr20", 57400000, 57400000 + 100)

# Bases, qualities, start position, length and name of first read
info$reads[[1]]
info$quality[[1]]
info$position[1]
info$lengths[1]
info$names[1]
}
\author{
Martin Vincent
}

