% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_C}
\alias{locate_C}
\title{Locate C}
\usage{
locate_C(object)
}
\arguments{
\item{object}{epiG object}
}
\value{
positions of C in object
}
\description{
Locate C positions
}
\examples{
data(example)

locate_C(fit)
}
\author{
Martin Vincent
}

