% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_CG}
\alias{locate_CG}
\title{Locate CpG}
\usage{
locate_CG(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
a vector of CpG positions
}
\description{
Locate CpG positions in the refrence genom
}
\examples{
data(example)

locate_CG(fit)
}
\author{
Martin Vincent
}

