% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_DGCH}
\alias{locate_DGCH}
\title{Locate DGCH}
\usage{
locate_DGCH(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
a vector of isolated GpC positions
}
\description{
Locate DGCH (isolated GpC) positions in the refrence genom
}
\examples{
data(example)

fit <- fetch_ref(fit)
locate_DGCH(fit)
}
\author{
Martin Vincent
}

