% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_DGCH}
\alias{locate_DGCH}
\title{Locate DGCH}
\usage{
locate_DGCH(object)
}
\arguments{
\item{object}{}
}
\value{
??
}
\description{
Locate DGCH (isolated GpC) positions
}
\examples{
#TODO examples locate_DGCH
}
\author{
Martin Vincent
}

