% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-locate.R
\name{locate_HCGD}
\alias{locate_HCGD}
\title{Locate HCGD}
\usage{
locate_HCGD(object)
}
\arguments{
\item{object}{epiG model}
}
\value{
a vector of isolated CpG positions
}
\description{
locate HCGD (isolated CpG) positions in the refrence genom
}
\examples{
#TODO examples locate_HCGD
}
\author{
Martin Vincent
}

