% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{nchain.epiG}
\alias{nchain.epiG}
\title{Number of chains}
\usage{
\method{nchain}{epiG}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
??
}
\description{
Number of chains
}
\author{
Martin Vincent
}

