% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{nchain}
\alias{nchain}
\title{Number of chains}
\usage{
nchain(object, pos, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}
}
\value{
??
}
\description{
Number of chains
}
\author{
Martin Vincent
}

