% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchain}
\alias{nchain}
\title{Number of chains}
\usage{
nchain(object, pos, ...)
}
\arguments{
\item{object}{epiG model}

\item{pos}{position}

\item{...}{additonal arguments}
}
\value{
??
}
\description{
Number of chains
}
\author{
Martin Vincent
}

