% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchunks.epiG}
\alias{nchunks.epiG}
\title{Number of chunks}
\usage{
\method{nchunks}{epiG}(object, ...)
}
\arguments{
\item{object}{epiG model}

\item{...}{ignored}
}
\value{
the number of chunks in the epiG object
}
\description{
Number of chunks in the epiG object
}
\examples{
#TODO examples nchunks
}
\author{
Martin Vincent
}

