% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nchunks}
\alias{nchunks}
\title{Number of chunks}
\usage{
nchunks(object, pos, ...)
}
\arguments{
\item{object}{an epiG model}

\item{pos}{position}

\item{...}{ignored}
}
\value{
??
}
\description{
Number of chunks
}
\author{
Martin Vincent
}

