% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{nread}
\alias{nread}
\title{Number of reads in model}
\usage{
nread(object, ...)
}
\arguments{
\item{object}{an epiG model}

\item{...}{additonal arguments}
}
\value{
number of reads in the model
}
\description{
Number of reads in model
}
\author{
Martin Vincent
}

