% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{position_info.epiG}
\alias{position_info.epiG}
\title{position_info}
\usage{
\method{position_info}{epiG}(object, pos, ...)
}
\arguments{
\item{object}{an epiG model}

\item{pos}{position}

\item{...}{ignored}
}
\value{
??
}
\description{
position_info
}
\examples{
#TODO
}
\author{
Martin Vincent
}

