% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{position_info}
\alias{position_info}
\title{Position information}
\usage{
position_info(object, pos, ...)
}
\arguments{
\item{object}{an epiG model}

\item{pos}{position}

\item{...}{additonal arguments}
}
\value{
a data.frame with \code{length(pos)} rows and the following columns
#TODO return position_info
}
\description{
Exstract information about positions in the model
}
\author{
Martin Vincent
}

