% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{position_info}
\alias{position_info}
\title{position_info}
\usage{
position_info(object, pos, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}
}
\value{
??
}
\description{
position_info
}
\author{
Martin Vincent
}

