% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG.config}
\alias{print.epiG.config}
\title{print config}
\usage{
\method{print}{epiG.config}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
print config
}
\author{
Martin Vincent
}

