% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG.model}
\alias{print.epiG.model}
\title{Print Information About an epiG Conversion Model}
\usage{
\method{print}{epiG.model}(x, ...)
}
\arguments{
\item{x}{epiG conversion model}

\item{...}{ignored}
}
\description{
Print Information About an epiG Conversion Model
}
\examples{
#TODO examples print.epiG.model 
}
\author{
Martin Vincent
}

