% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-summary.R
\name{print.epiG}
\alias{print.epiG}
\title{Print Information About an Fitted epiG Model}
\usage{
\method{print}{epiG}(x, ...)
}
\arguments{
\item{x}{epiG model}

\item{...}{ignored}
}
\description{
Print Information About an Fitted epiG Model
}
\author{
Martin Vincent
}

