% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_depth}
\alias{read_depth}
\title{read_depth}
\usage{
read_depth(object, pos = NULL, ...)
}
\arguments{
\item{object}{}

\item{pos}{}

\item{...}{}
}
\value{
??
}
\description{
read_depth
}
\author{
Martin Vincent
}

