% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_depth}
\alias{read_depth}
\title{Read depth}
\usage{
read_depth(object, pos = NULL)
}
\arguments{
\item{object}{epiG model}

\item{pos}{position}
}
\value{
??
}
\description{
Retrive the read depth at a given position in the model
}
\examples{
#TODO examples read_depth
}
\author{
Martin Vincent
}

