% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-fetch.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA File}
\usage{
read_fasta(file, refname, start, len, offset = 1)
}
\arguments{
\item{file}{path to fasta file}

\item{refname}{reference name}

\item{start}{start of region}

\item{len}{length of region}

\item{offset}{file offset (position of first base in file, usually offset = 1)}
}
\value{
a vector of length \code{len} containing the bases
}
\description{
Read a raw FASTA file
}
\examples{
ref_file <- system.file("extdata", "hg19_GNAS.fa", package="epiG")

# Specify region
chr <- "chr20"
start <- 57400000 
end <- 57400000 + 100

#Note that usually offset = 1
read_fasta(ref_file, chr, start, len = end-start+1, offset = 57380000)
}
\author{
Martin Vincent
}

