% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_fetch.R
\name{read_fasta}
\alias{read_fasta}
\title{Read fasta}
\usage{
read_fasta(filename, refname, start, len)
}
\arguments{
\item{filename}{}

\item{refname}{}

\item{start}{}

\item{len}{}
}
\value{
??
}
\description{
Read fasta
}
\author{
Martin Vincent
}

