% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_info.epiG}
\alias{read_info.epiG}
\title{read_info}
\usage{
\method{read_info}{epiG}(object, inc.symbols = FALSE, ...)
}
\arguments{
\item{object}{}

\item{inc.symbols}{}

\item{...}{}
}
\value{
??
}
\description{
read_info
}
\author{
Martin Vincent
}

