% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{read_info.epiG_reads}
\alias{read_info.epiG_reads}
\title{Information about reads}
\usage{
\method{read_info}{epiG_reads}(object, inc.symbols = FALSE, ...)
}
\arguments{
\item{object}{epiG read object}

\item{inc.symbols}{if \code{TRUE} each line in the returned data.frame will correspond to one nuleobase, if \code{FALSE} one read.}

\item{...}{ignored}
}
\value{
??
}
\description{
Retrive information about reads from a epiG read object produced with \code{load_reads}
}
\examples{
#TODO examples read_info
}
\author{
Martin Vincent
}

