% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{start.epiG}
\alias{start.epiG}
\title{start position}
\usage{
\method{start}{epiG}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
numeric
}
\description{
start position
}
\examples{
#TODO
}
\author{
Martin Vincent
}

