% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{start.epiG}
\alias{start.epiG}
\title{Start position}
\usage{
\method{start}{epiG}(object, ...)
}
\arguments{
\item{object}{an epiG model}

\item{...}{ignored}
}
\value{
the first position in the model
}
\description{
Return the first position in the model
}
\examples{
#TODO examples start.epiG
}
\author{
Martin Vincent
}

