% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{start}
\alias{start}
\title{Start position}
\usage{
start(object, ...)
}
\arguments{
\item{object}{an epiG model}

\item{...}{additonal arguments}
}
\value{
the first position in the model
}
\description{
Return the first position in the model
}
\author{
Martin Vincent
}

