\name{subregion.epiG}
\alias{subregion.epiG}
\title{subregion}
\usage{
  \method{subregion}{epiG} (object, start, end,
    chop.reads = TRUE, ...)
}
\arguments{
  \item{object}{}

  \item{start}{}

  \item{end}{}

  \item{chop.reads}{}

  \item{...}{}
}
\value{
  ??
}
\description{
  subregion
}
\author{
  Martin Vincent
}

