% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_navigate.R
\name{subregion.epiG}
\alias{subregion.epiG}
\title{subregion}
\usage{
\method{subregion}{epiG}(object, start, end, chop.reads = FALSE, ...)
}
\arguments{
\item{object}{}

\item{start}{}

\item{end}{}

\item{chop.reads}{}

\item{...}{}
}
\value{
??
}
\description{
subregion
}
\author{
Martin Vincent
}

