% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{subregion}
\alias{subregion}
\title{Subregion}
\usage{
subregion(object, start, end, chop.reads = FALSE, ...)
}
\arguments{
\item{object}{epiG model}

\item{start}{start of subregion}

\item{end}{end of subregion}

\item{chop.reads}{if TRUE reads will be choped at the boundaries of the region}

\item{...}{additonal arguments}
}
\value{
an epiG model
}
\description{
Cut out a subregion of the model
}
\author{
Martin Vincent
}

