% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG-navigate.R
\name{subregion}
\alias{subregion}
\title{Subregion}
\usage{
subregion(object, start, end, chop.reads = FALSE)
}
\arguments{
\item{object}{epiG model}

\item{start}{start position of subregion}

\item{end}{end position of subregion}

\item{chop.reads}{if TRUE reads will be choped at the boundaries of the region}
}
\value{
an epiG model
}
\description{
Exatract a subregion of an epiG model
}
\examples{
#TODO examples subregion
}
\author{
Martin Vincent
}

