% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiG_locate.R
\name{vector_search}
\alias{vector_search}
\title{vector_search}
\usage{
vector_search(pattern, text)
}
\arguments{
\item{pattern}{}

\item{text}{}
}
\value{
??
}
\description{
vector_search
}
\author{
Martin Vincent
}

